Fix some format specifier warnings
The warnings were produced by GCC 9.2.x for x86_64-linux-gnu or
i386-pc-msdosdjgpp targets.

Most of the fixes involve changing the type of a variable rather than
the format specifier. For many of the affected test conuter variables,
a basic int seems sufficient.

Some format specifier warnings still remain for cases where changing
type or casting seemed inappropriate. Those warnings will probably
require some new format specifier macros (e.g. SDL_PRIu32).

+++ test/testautomation_sdltest.c
@@ -1163,7 +1163,7 @@ sdltest_randomAsciiStringWithMaximumLength(void *arg)
 
   targetLen = 16 + SDLTest_RandomUint8();
   result = SDLTest_RandomAsciiStringWithMaximumLength((int) targetLen);
-  SDLTest_AssertPass("Call to SDLTest_RandomAsciiStringWithMaximumLength(%d)", targetLen);
+  SDLTest_AssertPass("Call to SDLTest_RandomAsciiStringWithMaximumLength(%d)", (int) targetLen);
   SDLTest_AssertCheck(result != NULL, "Validate that result is not NULL");
   if (result != NULL) {
      len = SDL_strlen(result);
@@ -1184,7 +1184,7 @@ sdltest_randomAsciiStringWithMaximumLength(void *arg)
   /* Negative test */
   targetLen = 0;
   result = SDLTest_RandomAsciiStringWithMaximumLength((int) targetLen);
-  SDLTest_AssertPass("Call to SDLTest_RandomAsciiStringWithMaximumLength(%d)", targetLen);
+  SDLTest_AssertPass("Call to SDLTest_RandomAsciiStringWithMaximumLength(%d)", (int) targetLen);
   SDLTest_AssertCheck(result == NULL, "Validate that result is NULL");
   lastError = (char *)SDL_GetError();
   SDLTest_AssertPass("SDL_GetError()");
@@ -1217,7 +1217,7 @@ sdltest_randomAsciiStringOfSize(void *arg)
   /* Positive test */
   targetLen = 16 + SDLTest_RandomUint8();
   result = SDLTest_RandomAsciiStringOfSize((int) targetLen);
-  SDLTest_AssertPass("Call to SDLTest_RandomAsciiStringOfSize(%d)", targetLen);
+  SDLTest_AssertPass("Call to SDLTest_RandomAsciiStringOfSize(%d)", (int) targetLen);
   SDLTest_AssertCheck(result != NULL, "Validate that result is not NULL");
   if (result != NULL) {
      len = SDL_strlen(result);
@@ -1238,7 +1238,7 @@ sdltest_randomAsciiStringOfSize(void *arg)
   /* Negative test */
   targetLen = 0;
   result = SDLTest_RandomAsciiStringOfSize((int) targetLen);
-  SDLTest_AssertPass("Call to SDLTest_RandomAsciiStringOfSize(%d)", targetLen);
+  SDLTest_AssertPass("Call to SDLTest_RandomAsciiStringOfSize(%d)", (int) targetLen);
   SDLTest_AssertCheck(result == NULL, "Validate that result is NULL");
   lastError = (char *)SDL_GetError();
   SDLTest_AssertPass("SDL_GetError()");
